renaming main to base module
