+++ tests/test_initialization.py
@@ -3,19 +3,19 @@ from unittest import mock
 import pytest
 
 import up42
-from up42 import catalog, main, storage, tasking
+from up42 import base, catalog, storage, tasking
 
 from .fixtures import fixtures_globals as constants
 
 
 def test_fails_to_initialize_if_not_authenticated():
-    with pytest.raises(main.UserNotAuthenticated):
+    with pytest.raises(base.UserNotAuthenticated):
         up42.initialize_catalog()
-    with pytest.raises(main.UserNotAuthenticated):
+    with pytest.raises(base.UserNotAuthenticated):
         up42.initialize_storage()
-    with pytest.raises(main.UserNotAuthenticated):
+    with pytest.raises(base.UserNotAuthenticated):
         up42.initialize_order(order_id=constants.ORDER_ID)
-    with pytest.raises(main.UserNotAuthenticated):
+    with pytest.raises(base.UserNotAuthenticated):
         up42.initialize_asset(asset_id=constants.ASSET_ID)
 
 
@@ -24,7 +24,7 @@ def test_should_initialize_objects(
     order_mock,
     asset_mock,
 ):
-    with mock.patch("up42.main.workspace") as workspace_mock:
+    with mock.patch("up42.base.workspace") as workspace_mock:
         workspace_mock.id = constants.WORKSPACE_ID
         workspace_mock.auth = auth_mock
